requirements.txt
flask>=3.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.2
networkx>=3.0
PyJWT>=2.8
openpyxl>=3.1
hmmlearn>=0.3.0
ib_insync>=0.9.86
# Bloomberg BLPAPI (install separately — requires Bloomberg Terminal):
# pip install --index-url=https://blpapi.bloomberg.com/repository/releases/python/simple blpapi
